Refactor disc handling to emulator
